+++ protractor.conf.js
@@ -1,0 +1,4 @@
+exports.config = {
+    seleniumAddress: 'http://selenium:4444/wd/hub',
+    allScriptsTimeout: 30000 // extended timeout, initial load of the test webserver may take a while
+};